+++ CHANGELOG.md
@@ -6,6 +6,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Checkpoint] - 2026-02-14 10:12:01
+
+### Changed Files
+- `supabase/functions/slack-ip-leaderboard/index.ts`
+- `supabase/migrations/20260214100727_fix_ip_leaderboard_include_mtd.sql`
+
+### Statistics
+```
+ 2 files changed, 158 insertions(+), 19 deletions(-)
+```
+
 ## [Checkpoint] - 2026-02-14 09:49:43
 
 ### Changed Files